requirements.txt
django
djangorestframework
llama-cpp-python